refact(Transactions): refactor components to for less code and fix double loading
