+++ src/components/transactions/scene.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Link, Route, Switch } from 'react-router-dom';
 
 import RequireLogin from 'components/utils/require-login';
 import RequireOrganization from 'components/require-organization';
@@ -17,6 +17,8 @@ const TransactionsScene: React.SFC<{}> = () => (
           <Route exact path="/transactions/new" component={ New } />
           <Route exact path="/transactions/:id/edit" component={ Edit } />
         </Switch>
+        <Link to="/transactions/new" className="btn btn-default pull-right">New Transaction...</Link>
+        <h1>Transactions</h1>
         <List />
       </MainLayout>
     </RequireOrganization>
